pyproject.toml
[project]
name = "palette-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "uvicorn[standard]>=0.44.0",
]
